+++ firmware/shared/srcpp/lib/lib_assert.cpp
