+++ firmware/shared/srcpp/lib/lib_ubsan.cpp
