+++ firmware/shared/srcpp/lib/lib_ubsan.hpp
